+++ app/BUILD.bazel
@@ -683,6 +683,7 @@ TEST_DEPS = [
     "//testing",
     "//testing/src/main/java/org/oppia/android/testing/robolectric:is_on_robolectric",
     "//testing/src/main/java/org/oppia/android/testing/robolectric:test_module",
+    "//testing/src/main/java/org/oppia/android/testing/threading:coroutine_executor_service",
     "//testing/src/main/java/org/oppia/android/testing/threading:test_module",
     "//testing/src/main/java/org/oppia/android/testing/time:test_module",
     "//third_party:androidx_annotation_annotation",
