+++ testing/src/main/java/org/oppia/android/testing/threading/BUILD.bazel
@@ -78,6 +78,24 @@ kt_android_library(
     ],
 )
 
+# This package should have limited visibility only to the utilities that need it to discourage broad
+# use.
+kt_android_library(
+    name = "coroutine_executor_service",
+    testonly = True,
+    srcs = [
+        "CoroutineExecutorService.kt",
+    ],
+    visibility = [
+        "//app:__subpackages__",
+        "//testing/src/test/java/org/oppia/android/testing/threading:__pkg__",
+    ],
+    deps = [
+        "//third_party:androidx_annotation_annotation",
+        "//third_party:org_jetbrains_kotlinx_kotlinx-coroutines-core",
+    ],
+)
+
 kt_android_library(
     name = "test_module",
     testonly = True,
