+++ testing/src/test/java/org/oppia/android/testing/threading/BUILD.bazel
@@ -20,6 +20,23 @@ kt_android_library(
 )
 
 oppia_android_test(
+    name = "CoroutineExecutorServiceTest",
+    srcs = ["CoroutineExecutorServiceTest.kt"],
+    custom_package = "org.oppia.android.data.testing",
+    test_class = "org.oppia.android.testing.threading.CoroutineExecutorServiceTest",
+    test_manifest = "//testing/src/test:test_manifest",
+    deps = [
+        ":dagger",
+        "//testing",
+        "//testing/src/main/java/org/oppia/android/testing/robolectric:test_module",
+        "//testing/src/main/java/org/oppia/android/testing/threading:coroutine_executor_service",
+        "//testing/src/main/java/org/oppia/android/testing/threading:test_module",
+        "//testing/src/main/java/org/oppia/android/testing/time:test_module",
+        "//third_party:robolectric_android-all",
+    ],
+)
+
+oppia_android_test(
     name = "TestCoroutineDispatcherEspressoImplTest",
     srcs = ["TestCoroutineDispatcherEspressoImplTest.kt"],
     custom_package = "org.oppia.android.data.testing",
